chore: remove all pass-thru capabilities and fold the remaining survivors into the base widget.
